Starting to implement filler skipping.
Filler skipping will automatically skip over advertising and
station announcement filler between the programs.
This commit is a baby step that starts in that direction.
It never finds the filler, but proves the logic for giving up
and dumping the front of the audio when it fails.

Also made minor improvements to printing floating point values.