Fixed bug calculating correct sample position for partial chunk.
In GlitchRemover::Fix(), the position calculated for a partial
chunk that has just been filled was wrong.
Added sanity check to catch these cases.
Next I will remove the need to pass a position argument to
ProcessChunk, because each chunk knows its position.
Just checking this in so I capture the state where the bug is proven fixed.