First experimental algorithm for deleting filler between program hours.
Find the first silent gap after 55 seconds into the second..fourth hours.
Delete everything before the middle of this silent gap.
This is a pretty safe rule that won't accidentally delete part of the program,
but will often still leave in some filler.  It is not perfect, but a
step in the right direction.