Fixed #11 - Try for 59.0 minutes for Hearts of Space, 58.5 for Echoes.
Hearts of Space programs turn out to be very close to 59:00 in length,
while Echoes tend to be very close to 58:30.
If the filename indicates the fourth hour of date that is on
a Saturday, assume Hearts of Space and try to make the program
59:00 long.

Also fixed a bug calculating the exact unfiltered program length
and updating the program offset that were causing accuracy problems.
Now the error values are much smaller.