Detect glitches using specified headroom, rather than histogram analysis.
Before this change, the program used histogram analysis to pick
a headroom value that included all but 1/D of the samples, where
D was a large denominator value.

The problem was that the number of glitches is highly variable
from week to week. When there are many glitches, we can end up
leaving too many glitches in, and when there are few glitches,
we can chop out audio that has nothing wrong with it.

Now just use a headroom value specified in dB.
Can override the default using the new -h option.